When writing the diff file, Niffy does not wait for the close of the stream.
We encountered this issue while adding an upload to S3 to our after phase.
The file uploaded to S3 contained partial data; indicating that there was two process competing to write and read the file.
Fix is a simple on close handler.
We also bumped the version of pngjs.

note: potential bug/omission found in pngjs
https://github.com/lukeapage/pngjs/blob/master/lib/png.js#L43